move NoOp scheduler from registry.cpp to no_op.cpp
