+++ csrc/scheduler/registry.cpp
@@ -178,106 +178,6 @@ namespace {
 //!
 //! Typical use case of this scheduler is to handle edge cases
 //!  such as where all tensors are size-1 or size-0.
-class NoOpScheduler : public SchedulerEntry {
-  //! Provides a dummy heuristic type to ensure
-  //!  unified interface on NoOp scheduler.
-  class NoOpHeuristic : public HeuristicParams {
-   public:
-    using HeuristicParams::HeuristicParams;
-
-    size_t hash() const override {
-      return 0;
-    }
-    std::shared_ptr<HeuristicParams> clone() const override {
-      return std::make_shared<NoOpHeuristic>();
-    }
-    bool sameAs(const std::shared_ptr<HeuristicParams>& other) const override {
-      auto other_casted = std::dynamic_pointer_cast<ReductionParams>(other);
-      return other_casted != nullptr && other_casted->cparams == cparams;
-    };
-  };
-
- public:
-  explicit NoOpScheduler(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr)
-      : SchedulerEntry(ScheduleHeuristic::NoOp) {
-    params_ = std::make_shared<NoOpHeuristic>("", runtime_info.getIndexType());
-  }
-
-  //! Check if the no-op heuristics apply in given fusion
-  static bool canScheduleCompileTime(Fusion* fusion) {
-    if (fusion->isNoOp()) {
-      return true;
-    }
-    // Check there're no non-trivial reduction ops.
-    for (auto reduction : ir_utils::getReductionOps(fusion)) {
-      for (auto output :
-           ir_utils::filterByType<TensorView>(reduction->outputs())) {
-        auto concrete_dimension =
-            TensorDomain::noReductions(output->getRootDomain());
-        auto all_nonzero = std::none_of(
-            concrete_dimension.begin(),
-            concrete_dimension.end(),
-            [](IterDomain* id) { return id->extent()->isZeroInt(); });
-        if (all_nonzero) {
-          scheduler_debug_utils::canScheduleRejectReason(
-              ScheduleHeuristic::NoOp,
-              "reduction of non-zero elements is not supported");
-          return false;
-        }
-      }
-    }
-
-    // Check that all outputs are either broadcast or ignored reduction.
-    for (auto out_tv : ir_utils::filterByType<TensorView>(fusion->outputs())) {
-      auto concrete_dimension = TensorDomain::noReductions(
-          TensorDomain::noBroadcasts(out_tv->getLeafDomain()));
-      if (!concrete_dimension.empty()) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::NoOp, "output has a concrete dimension");
-        return false;
-      }
-    }
-
-    // Check that inputs of all select/gather-like ops are fusion inputs
-    if (registry_utils::rejectScheduleForMemoryPromotion(
-            fusion, ScheduleHeuristic::NoOp)) {
-      return false;
-    }
-
-    // We have verified that all iterdomains on all output tv's are trivial
-    // reductions,
-    //  broadcasts or zero-sized. Therefore accepting this fusion for NoOp
-    //  scheduling.
-    return true;
-  }
-
-  static bool canScheduleRunTime(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr) {
-    // TODO:
-    //  Pipe through dynamic zero checks.
-    return true;
-  }
-
-  void schedule(Fusion* fusion) override {
-    // Schedule is no-op.
-    return;
-  }
-
- private:
-  void computeHeuristics(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr) {
-    // Heuristics is no-op.
-    return;
-  }
-};
-
 class ReductionScheduler : public SchedulerEntry {
  public:
   explicit ReductionScheduler(
